+++ src/stores/Theme/index.state.ts
@@ -1,0 +1,13 @@
+const themeState = {
+  themeState: {
+    isDark: () => {
+      const result = !!(
+        window.matchMedia &&
+        window.matchMedia('(prefers-color-scheme: dark)').matches
+      );
+      return window !== undefined ? result : false;
+    },
+  },
+};
+
+export default themeState;