+++ src/utilities/useCategoryMainVisible/index.ts
@@ -1,0 +1,18 @@
+// import { useEffect } from 'react';
+
+/**
+ * TODO: wpgql 이라는 데이터로부터 시작한다.
+ * 이 데이터를 받아서 post타입이던 categoy, taxonomy 타입이던 아무거나 상관없이
+ * useCategoryMainVisible(wpgql) 후에 메인에 보여지지 않는 단독 카테고리만 가지고있다면
+ * 해당 배열을 제거후 원래 데이터를 리턴한다.
+ * 필터 개념
+ */
+
+// const isFilterCategoryReducer = (acc, cur) => {
+//   const {};
+// };
+// const isCategoryData = (data) => {
+//   const Keys = Object.keys(data);
+//   return !!Keys.map((key) => /categor[ies|y]/.test(key)) || false;
+// };
+// const useCategoryMainVisible = () => {};