+++ src/views/components/organisms/MSB/index.tsx
@@ -1,22 +1,24 @@
 import React, { useRef } from 'react';
 // Components
-import { Logo } from '@atom/Logo';
-import { MenuStateType } from '@store/rootStore';
+import Logo from '@atom/Logo';
+import { MenuStateType } from '@store/Menu';
 import useOutsideAlerter from '@utile/useOutsideAlerter';
+// Style
+import { SideBar, LogoLayout, Main, Row, Col } from './index.style';
 // Containers
 import DayAndNightToggle from '#/DayAndNight';
 import PostTypeAllCat from '#/PostTypeAllCat';
-// Style
-import { SideBar, LogoLayout, Main, Row, Col } from './index.style';
+import StaticAllPageList from '#/StaticAllPageList';
 
 interface props extends MenuStateType {
-  inFunction: any;
+  inFunction?: any;
+  type: string;
 }
-const MSB: React.FC<props> = ({ isActive, inFunction }: props) => {
+const MSB: React.FC<props> = ({ isActive, inFunction, type }: props) => {
   const wrapperRef = useRef<HTMLDivElement>(null);
   useOutsideAlerter({ ref: isActive && wrapperRef, inFunction });
   return (
-    <SideBar ref={wrapperRef} isActive={isActive}>
+    <SideBar ref={wrapperRef} isActive={isActive} type={type}>
       <LogoLayout>
         <Logo />
       </LogoLayout>
@@ -27,6 +29,15 @@ const MSB: React.FC<props> = ({ isActive, inFunction }: props) => {
           </Col.Def>
         </Row.Menu>
         <Row.Def>
+          <Col.Def col>
+            <div>Page</div>
+            <StaticAllPageList />
+          </Col.Def>
+        </Row.Def>
+        <Row.Def>
+          <Col.Def col>
+            <div>Category</div>
+          </Col.Def>
           <Col.Def col>
             <PostTypeAllCat />
           </Col.Def>