+++ src/views/styles/themes/index.ts
@@ -10,19 +10,19 @@ interface Colors {
   grayscales: Grayscales;
 }
 
-interface DarkColors extends Colors {}
+// interface DarkColors extends Colors {}
 interface StyleSystem {
   color: Colors;
 }
 
-interface DarkStyleSystem extends StyleSystem {
-  color: DarkColors;
-}
+// interface DarkStyleSystem extends StyleSystem {
+//   color: DarkColors;
+// }
 
 const DefaultTheme: StyleSystem = {
   color: {
     primary: '#0f4c81',
-    bg: ['#fff', '#f6f8ff'],
+    bg: ['#fff', '#E1EBF2'],
     text: ['#333', '#4A4F57', '#868E96', '#dcdcdc'],
     grayscales: {
       dark: ['#12171f', '#161a23', '#181c27', '#1f2330', '#252836', '#303645'],
@@ -31,12 +31,12 @@ const DefaultTheme: StyleSystem = {
   },
 };
 
-export const DarkTheme: DarkStyleSystem = {
+export const DarkTheme: StyleSystem = {
   color: {
     ...DefaultTheme.color,
     primary: '#fff46a',
-    bg: ['#303437', '#232326'],
-    text: ['#f2f4f7', '#e6ecf8', '#9ba4b7', '#515151'],
+    bg: ['#2C2F33', '#232326'],
+    text: ['#C2BFB9', '#C2BFB9', '#9ba4b7', '#515151'],
   },
 };