+++ CHANGELOG.md
@@ -3,6 +3,42 @@
 이 프로젝트에서 주목할만한 모든 변경 사항이이 파일에 문서화됩니다.
 [Keep a Changelog](https://keepachangelog.com/ko/1.0.0/)의 형식을 기본으로 구성됩니다. 현재 버전은 [#181](https://github.com/Hansanghyeon/4log/issues/181)위와 같은 내용으로 구성됩니다.
 
+## [1.13.0] - 2020-09-10
+
+### Added
+
+- Lottie 애니메이션중 [UnderMaintenance](https://lottiefiles.com/6873-under-maintenance) 추가
+
+### Changed
+
+- Theme Color 눈에 보기 편하게 색조정
+- About Page에서 푸터사라지는 것 수정 푸터사라지지 않게
+
+## [1.12.0] - 2020-09-10
+
+### Added
+
+- About Page 추가
+
+### Changed
+
+- About Page Redux 상태 적용
+- 패키지 최신화
+- react-hook에 맞게 타입들 변경
+
+## [1.11.0] - 2020-08-14
+
+### Changed
+
+- project 임시 페이지 완료 더이상 아이디어가 없어서 진행하지 못함
+- Redux 레이아웃 타입 추가
+
+## [1.10.3] - 2020-08-09
+
+### Fixed
+
+- `html-react-parser` 패키지가 코드블럭의 string으로 표현되어야할 태그들까지 모두 html tag로 변경해주는 것을 `parser`를 사용할때 `options`에다가 원본 string값을 넘겨서 노출되지않던 부분까지 모두 보이게 수정
+
 ## [1.10.2] - 2020-07-05
 
 ### Fixed